requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.1
Django==5.0.4
django-crispy-forms==2.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
gunicorn==22.0.0
mysqlclient==2.2.4
packaging==24.0
pillow==10.3.0
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.6.0
